add id check

+++ server/src/controllers/user.controller.js
@@ -77,17 +77,23 @@ const userController = {
   async deactivate(req, res, next) {
     const transaction = await sequelize.transaction();
     try {
-      const user = await User.findByPk(req.params.id);
-      await user.destroy({ transaction });
-      await transaction.commit();
-      res.status(200).json({
-        message: "The membership withdrawal was completed."
-      });
+      if (Number(req.params.id) === req.decoded.id) {
+        const user = await User.findByPk(req.params.id);
+        await user.destroy({ transaction });
+        await transaction.commit();
+        res.status(200).json({
+          message: "The membership withdrawal was completed."
+        });
+      } else {
+        return next(boom.badRequest("A bad request was sent."));
+      }
     } catch (error) {
       await transaction.rollback();
+
       error = boom.boomify(error);
       error.output.payload.message =
         "Sorry, our service is temporaily unavailable.";
+
       next(error);
     }
   }
